SettingDto - set mileage int -> double

+++ CA.Ticketing.Business/Services/Settings/Dto/SettingDto.cs
@@ -6,6 +6,6 @@
 
         public double OvertimePercentageIncrease { get; set; }
 
-        public int MileageCost { get; set; }
+        public double MileageCost { get; set; }
     }
 }
